refactor(cockpit): move contract tests to owner module
