+++ crates/tokmd-cockpit/src/contracts.rs
@@ -38,3 +38,58 @@ pub fn detect_contracts<S: AsRef<str>>(files: &[S]) -> Contracts {
         breaking_indicators,
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_detect_contracts_api() {
+        let files = vec!["crates/tokmd-types/src/lib.rs"];
+        let contracts = detect_contracts(&files);
+        assert!(contracts.api_changed);
+        assert!(!contracts.cli_changed);
+        assert!(!contracts.schema_changed);
+        assert_eq!(contracts.breaking_indicators, 1);
+    }
+
+    #[test]
+    fn test_detect_contracts_cli() {
+        let files = vec!["crates/tokmd/src/commands/lang.rs"];
+        let contracts = detect_contracts(&files);
+        assert!(!contracts.api_changed);
+        assert!(contracts.cli_changed);
+    }
+
+    #[test]
+    fn test_detect_contracts_schema() {
+        let files = vec!["docs/schema.json"];
+        let contracts = detect_contracts(&files);
+        assert!(contracts.schema_changed);
+        assert_eq!(contracts.breaking_indicators, 1);
+    }
+
+    #[test]
+    fn test_detect_contracts_none() {
+        let files = vec!["README.md", "src/utils.rs"];
+        let contracts = detect_contracts(&files);
+        assert!(!contracts.api_changed);
+        assert!(!contracts.cli_changed);
+        assert!(!contracts.schema_changed);
+        assert_eq!(contracts.breaking_indicators, 0);
+    }
+
+    #[test]
+    fn test_detect_contracts_all() {
+        let files = vec![
+            "crates/tokmd-types/src/lib.rs",
+            "crates/tokmd/src/commands/lang.rs",
+            "docs/schema.json",
+        ];
+        let contracts = detect_contracts(&files);
+        assert!(contracts.api_changed);
+        assert!(contracts.cli_changed);
+        assert!(contracts.schema_changed);
+        assert_eq!(contracts.breaking_indicators, 2); // api + schema
+    }
+}
